--- src/chains/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './backbone.entity';
