+++ src/datasources/transaction-api/entities/schemas/backbone.schema.ts
@@ -1,5 +1,5 @@
 import { JSONSchemaType } from 'ajv';
-import { Backbone } from '../../../../chains/entities/backbone.entity';
+import { Backbone } from '../backbone.entity';
 
 const backboneSchema: JSONSchemaType<Backbone> = {
   type: 'object',
