+++ src/datasources/transaction-api/transaction-api.service.spec.ts
@@ -1,12 +1,12 @@
 import { Balance } from './entities/balance.entity';
 import { TransactionApi } from './transaction-api.service';
-import { Backbone } from '../../chains/entities';
 import { CacheFirstDataSource } from '../cache/cache.first.data.source';
 import { ValidationErrorFactory } from '../errors/validation-error-factory';
 import { HttpException } from '@nestjs/common';
 import backboneFactory from '../../chains/entities/__tests__/backbone.factory';
 import { balanceFactory } from './entities/__tests__/balance.factory';
 import { JsonSchemaService } from '../../common/schemas/json-schema.service';
+import { Backbone } from './entities/backbone.entity';
 
 const BALANCES: Balance[] = [balanceFactory(), balanceFactory()];
 const BACKBONE: Backbone = backboneFactory();
